Comment out PFP4 entry in winwing_mcdu.py
I don't have the PFP4 device so I cannot check but it looks like there
is a PID error either for the PFP4 or the FCU + EFIS-L combo. They both
share the 0xbc1d PID in winwing_mcdu.py and winwing_fcu.py. This makes
it so an FCU + EFIS-L combo is recognized as an MCDU and then the script
triggers an exception when winwing_fcu.py tries to connect to the
device. This issue is not visible when both an MCDU and the FCU + EFIS-L
combo are connected since the MCDU is recognized before going over the
PFP4 entry.

Signed-off-by: Jeremie Corbier <jeremie.corbier@gmail.com>